Add files via upload
Changed doImport() to actually work right. Works on Guhem's, works on our exported, and now even works on a full config file.
Changed attackButton() to be selectable - pass it true or false to turn it on or off, or leave it nil to toggle.
Changed initial battlerage parsing to spam send(" ") less and still work perfectly.